Move filters to separated file
